Updated Company Class - updated spreading formula to be more based on security

+++ WIU-Project/Company.h
@@ -21,7 +21,7 @@ public:
 	~Company();
 
 	// Getters and Setters
-	int getInfectedStatus() const;
+	float getInfectedStatus() const;
 	int getNoOfInfectedComputers() const;
 	std::string getName() const;
 	int getNetworkSize() const;
